removed db-get-... and changed find to findOne, and test is returning res.status with error code
