--- spots-api/api/db-get-spot.js
@@ -1,11 +1,0 @@
-// db-get-spot.js
-module.exports = function(db) {
-  return function(req, res) {
-    var spots = db.collections('spots');
-    spots.find({_id:req.params.id}).toArray((err, result) => {
-      if(err) throw err;
-
-      res.json(result[0]);
-    });
-  };
-};
