+++ spots-api/api/get-spot.spec.js
@@ -2,13 +2,14 @@ const app = require('../app.js');
 const should = require('should');
 const supertest = require('supertest');
 
-describe('/spots/:id', (req, res) => {
-  it('should return a json object', done => {
+describe('/spots/:id', () => {
+  it('should return a spot from db', (done) => {
     supertest(app)
       .get('/spots/1')
-      .end(err, res) => {
-        res.body.id.should.equal('1');
+      .end((err, res) => {
+        if(err) throw err;
+        res.body.should.be.an.Object;
         done();
-      }
-  })
+      });
+  });
 });
